+++ src/ng2-page-scroll-config.ts
@@ -1,4 +1,4 @@
-export interface IEasingFunction {
+export abstract class EasingLogic {
     /**
      * Examples may be found at https://github.com/gdsmith/jquery.easing/blob/master/jquery.easing.js
      * or http://gizma.com/easing/
@@ -7,8 +7,12 @@ export interface IEasingFunction {
      * @param c change In value
      * @param d duration
      */
-    (t: number, b: number, c: number, d: number): number;
+    public abstract ease(t: number, b: number, c: number, d: number): number;
 }
+
+export declare type PageScrollTarget = HTMLElement | string;
+
+export declare type PageScrollingViews = HTMLElement | Document | HTMLBodyElement | Node;
 
 export class PageScrollConfig {
 
@@ -19,6 +23,8 @@ export class PageScrollConfig {
      * @private
      */
     public static _interval: number = 10;
+
+    public static _defaultNamespace = 'default';
 
     /**
      * The duration how long a scrollTo animation should last by default.
@@ -35,7 +41,7 @@ export class PageScrollConfig {
     public static defaultScrollOffset: number = 0;
 
     /**
-     * The events that are listened to on the body to decide whether a scroll animation has been interfered by the user
+     * The events that are listened to on the body to decide whether a scroll animation has been interfered/interrupted by the user
      * @type {string[]}
      * @private
      */
@@ -56,17 +62,20 @@ export class PageScrollConfig {
      */
     public static defaultInterruptible: boolean = true;
 
-    private static _easingFunction: IEasingFunction = (t: number, b: number, c: number, d: number): number => {
-        // Linear easing
-        return c * t / d + b;
+    private static _easingLogic: EasingLogic = {
+        ease: (t: number, b: number, c: number, d: number): number => {
+            // Linear easing
+            return c * t / d + b;
+        }
     };
 
     // Getter and setter to avoid auto completion to suggest calling the method
-    public static get defaultEasingFunction(): IEasingFunction {
-        return PageScrollConfig._easingFunction;
+    public static get defaultEasingLogic(): EasingLogic {
+        return PageScrollConfig._easingLogic;
     }
 
-    public static set defaultEasingFunction(easingFunction: IEasingFunction) {
-        PageScrollConfig._easingFunction = easingFunction;
+    public static set defaultEasingLogic(easingLogic: EasingLogic) {
+        PageScrollConfig._easingLogic = easingLogic;
     }
+
 }